Agrego Funcionalidad a la pestaña de reservas
- Capturo los valores de los inputs del formulario de reservas
-los guardo en un objeto booking
-guardo el objeto booking en la tabla de Booking en la base de datos

+++ TPCuatrimestral-Equipo-16/CabDominio/Booking.cs
@@ -15,7 +15,7 @@ namespace CabDominio
         public City Destination { get; set; } //Composición de ciudad
         public DateTime SolicitudDate { get; set; }
         public DateTime DateBooking { get; set; }
-        public int Passengers { get; set; }
+        public short Passengers { get; set; }
         public string StateBooking { get; set; } 
         public bool State { get; set; }
